Pairs Tests: * Move Package Structs from Numbers Module to Pairs Module * Move TimeMeasurement to Primes Test Package
