+++ pairs/src/test/java/mathtools/pairs/DoublePairFixedTest.java
@@ -1,4 +1,4 @@
-package mathtools.numbers.structs;
+package mathtools.pairs;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
@@ -6,26 +6,19 @@ import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-/** Testing [FloatPairFixed] class.
+/** Testing [DoublePairFixed] class.
  * @author DK96-OS : 2023
  */
-public final class FloatPairFixedTest {
+public final class DoublePairFixedTest {
 
-	private final float number1 = 1f;
-	private final float number2 = Float.MAX_VALUE;
+	private final double number1 = 1.0;
+	private final double number2 = Double.MAX_VALUE;
 
-	@Test
-	void testConstructor() {
-		final DoublePairFixed pair = new DoublePairFixed(number1, number2);
-		assertEquals(number1, pair.first);
-		assertEquals(number2, pair.second);
-	}
-
-	private FloatPairFixed mPair;
-
+	private DoublePairFixed mPair;
+	
 	@BeforeEach
 	void testSetup() {
-		mPair = new FloatPairFixed(number1, number2);
+		mPair = new DoublePairFixed(number1, number2);
 	}
 
 	@Test
@@ -36,40 +29,40 @@ public final class FloatPairFixedTest {
 
 	@Test
 	void testToMutable() {
-		FloatPair pair = mPair.toMutable();
+		DoublePair pair = mPair.toMutable();
 		assertEquals(mPair.first, pair.getFirst());
 		assertEquals(mPair.second, pair.getSecond());
 	}
 
 	@Test
 	void testEquals_SameValues_ReturnsTrue() {
-		FloatPairFixed pair = new FloatPairFixed(number1, number2);
+		DoublePairFixed pair = new DoublePairFixed(number1, number2);
 		assertEquals(mPair, pair);
 	}
 
 	@Test
 	void testEquals_DifferentValues_ReturnsFalse() {
-		FloatPairFixed pair = new FloatPairFixed(
+		DoublePairFixed pair = new DoublePairFixed(
 			number1, (byte) (1 + number2)
 		);
 		assertNotEquals(mPair, pair);
-		pair = new FloatPairFixed(
+		pair = new DoublePairFixed(
 			(byte) (1 + number1), number2
 		);
 		assertNotEquals(mPair, pair);
 	}
 
 	@Test
-	void testEquals_WithMutableFloatPair_ReturnsTrue() {
-		FloatPair pair0 = mPair.toMutable();
+	void testEquals_WithMutableDoublePair_ReturnsTrue() {
+		DoublePair pair0 = mPair.toMutable();
 		assertEquals(mPair, pair0);
 	}
 
 	@Test
-	void testEquals_WithMutableFloatPairDifferentValues_ReturnsFalse() {
-		FloatPair pair = new FloatPair(number1, (byte) 10);
+	void testEquals_WithMutableDoublePairDifferentValues_ReturnsFalse() {
+		DoublePair pair = new DoublePair(number1, (byte) 10);
 		assertNotEquals(mPair, pair);
-		pair = new FloatPair((byte) 10, number2);
+		pair = new DoublePair((byte) 10, number2);
 		assertNotEquals(mPair, pair);
 	}
 
@@ -82,7 +75,7 @@ public final class FloatPairFixedTest {
 	@Test
 	void testToString() {
 		assertEquals(
-			"(1.0, " + Float.MAX_VALUE + ")",
+			"(1.0, " + Double.MAX_VALUE + ")",
 			mPair.toString()
 		);
 	}
